Add a stress command to exercize the cgo wrappers.
By default, it runs 100 instances of jsonnet in parallel,
with lots of jsonnet operations in each, and does this all
in a loop 1000 times.  Flags can adjust these numbers.

This does demonstrate stability of the wrapper.

It could also demonstrate memory leaks,
but I have not yet run this long enough to really know
if we still have leaks.